refactor: Move UserRepository to repository package and update imports
