+++ src/test/java/com/aidom/api/domain/auth/service/AuthServiceIntegrationTest.java
@@ -8,7 +8,7 @@ import com.aidom.api.domain.auth.entity.AuthCode;
 import com.aidom.api.domain.auth.entity.RefreshToken;
 import com.aidom.api.domain.auth.repository.AuthCodeRepository;
 import com.aidom.api.domain.auth.repository.RefreshTokenRepository;
-import com.aidom.api.domain.user.UserRepository;
+import com.aidom.api.domain.user.repository.UserRepository;
 import com.aidom.api.domain.user.entity.User;
 import com.aidom.api.domain.user.enums.Provider;
 import com.aidom.api.domain.user.enums.Role;
